requirements.txt
alembic==1.0.1
banal==0.3.8
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dash==0.28.5
dash-core-components==0.34.0
dash-html-components==0.13.2
dash-renderer==0.14.3
dataset==1.1.0
decorator==4.3.0
facebook-business==3.1.2
-e git+https://github.com/mobolic/facebook-sdk.git@e997998838b547b27ca1dbbc3fce11a5638f8bac#egg=facebook_sdk
facebookads==2.11.4
Flask==1.0.2
Flask-Compress==1.4.0
gunicorn==19.9.0
idna==2.7
ipython-genutils==0.2.0
ItsDangerous==1.0.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.0
nbformat==4.4.0
normality==0.6.1
numpy==1.15.2
pandas==0.23.4
plotly==3.3.0
pydeepl==0.13
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.12
traitlets==4.3.2
urllib3==1.23
vaderSentiment==3.2.1
Werkzeug==0.14.1
